digit/requirements.txt
streamlit
numpy
tensorflow
streamlit-drawable-canvas
Pillow
opencv-python-headless
